split up systems
